Added minimal support for channelized OSC messages
Some OSC clients like f.e. TC-Data (https://bitshapesoftware.com/instruments/tc-data/)
have the concept of OSC channels. These channels are used like MIDI channels.

One example could be the first touching finger gets channel 1, the second channel 2 and so on.

An OSC message looks like this:  /device/layer/1/send/1/volume [ 371.29318, 1.0 ]

The second value in the data array is the channel (0.0 is first channel).

The idea is to re-write the layer id with the channel. Then in Bitwig the different fingers/channels
in TC-Data can be programmed with "Map to controller key" as different entities in global user macros.

A better solution would be to fully integrate the concept of the OSC result data array for different purposes.
Also handling of unique OSC strings like /global/metaparameter/{1-128} would be cleaner than to misuse the
/device/layer/. .. string for mapped user parameters.