Preliminary support for track/bank switching on parameter change
Sending the `/device/param/{1-8}/value` OSC message only affects the currently selected track and device. This commit adds preliminary support for a new OSC message, `/track/n/params/{1-8}/{1-8}/value`. Send the track number to be selected as n, then the parameter page and then the parameter number, and when the value is changed the specified track will be selected, the required parameter bank switch to, and the specified parameter's value will be changed. Tested in Bitwig 4 for Mac and TouchOSC on an iPad.

To do: feedback to TouchOSC when value changed in DAW